fix(workspace): address code review findings for VastuChart [VASTU-1B-135]
Fixes 2 red findings from PR review on #270:

1. Duplicate left legend: changed outer condition from
   `(legendPosition === 'top' || legendPosition === 'left')` to
   `legendPosition === 'top'` — left legend now only renders
   inside the flex container where it participates in horizontal layout.

2. Component line count: extracted 6 chart renderers and shared
   axis/grid/tooltip elements into dedicated files:
   - ChartRenderer.tsx: switch-based renderer component (~210 lines)
   - useChartElements.tsx: shared axis/grid/tooltip hook (~120 lines)
   Main VastuChart.tsx is now ~280 lines.

Also addresses yellow finding:
- prefersReducedMotion: replaced synchronous matchMedia call with
  useSyncExternalStore + subscribePrefersReducedMotion so the component
  reactively updates when the user toggles accessibility settings.
- ReferenceLine key changed from array index to composite key
  (rl-x-y-label) to avoid key instability on reorder/remove.
