+++ packages/workspace/src/components/VastuChart/VastuChart.tsx
@@ -19,35 +19,17 @@
  * Implements US-135 AC-1 through AC-12.
  */
 
-import React, { useState, useCallback, useMemo } from 'react';
+import React, { useState, useCallback, useMemo, useEffect, useSyncExternalStore } from 'react';
 import { Skeleton, Button } from '@mantine/core';
-import {
-  ResponsiveContainer,
-  LineChart,
-  Line,
-  BarChart,
-  Bar,
-  AreaChart,
-  Area,
-  PieChart,
-  Pie,
-  Cell,
-  ScatterChart,
-  Scatter,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-  ReferenceLine,
-  Legend as RechartsLegend,
-} from 'recharts';
+import { ResponsiveContainer, Legend as RechartsLegend } from 'recharts';
 import { IconAlertCircle, IconSettings, IconChartLine } from '@tabler/icons-react';
 
-import type { VastuChartProps, ChartConfig, SeriesConfig, ChartDataPoint } from './types';
+import type { VastuChartProps, ChartConfig, SeriesConfig } from './types';
 import { getSeriesColor } from './chartColors';
-import { ChartTooltip } from './ChartTooltip';
 import { ChartLegend } from './ChartLegend';
 import { ChartConfigPanel } from './ChartConfigPanel';
+import { ChartRenderer } from './ChartRenderer';
+import { useChartElements } from './useChartElements';
 import { EmptyState } from '../EmptyState';
 import { t } from '../../lib/i18n';
 import classes from './VastuChart.module.css';
@@ -59,18 +41,6 @@ const DEFAULT_HEIGHT = 240;
 
 /** Default chart height for full-width usage. */
 const DEFAULT_HEIGHT_FULL = 360;
-
-/** Donut inner radius ratio relative to outer. */
-const DONUT_INNER_RADIUS_RATIO = 0.6;
-
-/** Area fill opacity per SS10.5. */
-const AREA_FILL_OPACITY = 0.08;
-
-/** Active dot radius per SS10.5. Dots are hidden by default (r:0), shown on hover. */
-const ACTIVE_DOT_RADIUS = 5;
-
-/** Bar corner radius per SS10.5. */
-const BAR_CORNER_RADIUS = 2;
 
 /** Line stroke width for <= 5 series. */
 const LINE_STROKE_WIDTH = 2;
@@ -93,22 +63,37 @@ function resolveSeriesColors(series: SeriesConfig[]): Record<string, string> {
 }
 
 /**
- * Check if prefers-reduced-motion is active.
- * Returns true in SSR (conservative default).
+ * Subscribe to prefers-reduced-motion media query changes.
+ * Used by useSyncExternalStore to react when the user toggles accessibility settings.
  */
-function prefersReducedMotion(): boolean {
+function subscribePrefersReducedMotion(callback: () => void): () => void {
+  if (typeof window === 'undefined') return () => undefined;
+  const mq = window.matchMedia('(prefers-reduced-motion: reduce)');
+  mq.addEventListener('change', callback);
+  return () => mq.removeEventListener('change', callback);
+}
+
+function getPrefersReducedMotionSnapshot(): boolean {
   if (typeof window === 'undefined') return true;
   return window.matchMedia('(prefers-reduced-motion: reduce)').matches;
 }
 
-// ─── Shared axis/grid props ────────────────────────────────────────────────
+/** Server snapshot: conservative default returns true (no animation during SSR). */
+function getPrefersReducedMotionServerSnapshot(): boolean {
+  return true;
+}
 
-const AXIS_TICK_STYLE = {
-  fontFamily: 'var(--v-font-sans)',
-  fontSize: 11, // --v-text-xs
-  fill: 'var(--v-text-tertiary)',
-  fontWeight: 400,
-};
+/**
+ * Hook: reactively returns the current prefers-reduced-motion value.
+ * Updates immediately when the user toggles accessibility settings.
+ */
+function usePrefersReducedMotion(): boolean {
+  return useSyncExternalStore(
+    subscribePrefersReducedMotion,
+    getPrefersReducedMotionSnapshot,
+    getPrefersReducedMotionServerSnapshot,
+  );
+}
 
 // ─── Component ───────────────────────────────────────────────────────────────
 
@@ -132,13 +117,12 @@ export function VastuChart({
   const [internalConfig, setInternalConfig] = useState<ChartConfig>(config);
 
   // Sync internalConfig when prop changes
-  React.useEffect(() => {
+  useEffect(() => {
     setInternalConfig(config);
   }, [config]);
 
   const resolvedColors = useMemo(() => resolveSeriesColors(series), [series]);
-
-  const reducedMotion = prefersReducedMotion();
+  const reducedMotion = usePrefersReducedMotion();
 
   // Effective config (merge defaults)
   const effectiveConfig: Required<ChartConfig> = {
@@ -163,6 +147,15 @@ export function VastuChart({
 
   const strokeWidth = series.length > 5 ? LINE_STROKE_WIDTH_DENSE : LINE_STROKE_WIDTH;
 
+  // ─── Chart elements (shared axis/grid/tooltip) ─────────────────────────────
+
+  const chartElements = useChartElements({
+    effectiveConfig,
+    series,
+    resolvedColors,
+    reducedMotion,
+  });
+
   // ─── Legend handlers ───────────────────────────────────────────────────────
 
   const handleLegendToggle = useCallback((dataKey: string) => {
@@ -179,7 +172,6 @@ export function VastuChart({
 
   const handleLegendSolo = useCallback(
     (dataKey: string) => {
-      // If this series is currently the only visible one, restore all
       const allOtherKeys = series
         .filter((s) => s.dataKey !== dataKey)
         .map((s) => s.dataKey);
@@ -190,11 +182,9 @@ export function VastuChart({
         allOtherKeys.every((k) => currentlyHidden.has(k));
 
       if (isAlreadySoloed && previousHidden !== null) {
-        // Restore previous hidden state
         setHiddenSeries(previousHidden);
         setPreviousHidden(null);
       } else {
-        // Save current hidden state and solo this series
         setPreviousHidden(new Set(currentlyHidden));
         setHiddenSeries(new Set(allOtherKeys));
       }
@@ -282,285 +272,6 @@ export function VastuChart({
     );
   }
 
-  // ─── Shared chart children ────────────────────────────────────────────────
-
-  const cartesianGridEl = (
-    <CartesianGrid
-      strokeDasharray="3 3"
-      stroke="var(--v-border-subtle)"
-      vertical={false}
-    />
-  );
-
-  const xAxisEl = effectiveConfig.showAxisLabels ? (
-    <XAxis
-      dataKey={effectiveConfig.xAxisKey}
-      tick={AXIS_TICK_STYLE}
-      axisLine={{ stroke: 'var(--v-border-default)' }}
-      tickLine={false}
-      label={
-        effectiveConfig.xAxisLabel
-          ? {
-              value: effectiveConfig.xAxisLabel,
-              position: 'insideBottom',
-              offset: -4,
-              style: { ...AXIS_TICK_STYLE, fill: 'var(--v-text-secondary)' },
-            }
-          : undefined
-      }
-    />
-  ) : (
-    <XAxis dataKey={effectiveConfig.xAxisKey} hide />
-  );
-
-  const yAxisEl = effectiveConfig.showAxisLabels ? (
-    <YAxis
-      scale={effectiveConfig.scaleType === 'log' ? 'log' : 'auto'}
-      tick={AXIS_TICK_STYLE}
-      axisLine={{ stroke: 'var(--v-border-default)' }}
-      tickLine={false}
-      width={40}
-      label={
-        effectiveConfig.yAxisLabel
-          ? {
-              value: effectiveConfig.yAxisLabel,
-              angle: -90,
-              position: 'insideLeft',
-              style: { ...AXIS_TICK_STYLE, fill: 'var(--v-text-secondary)' },
-            }
-          : undefined
-      }
-    />
-  ) : (
-    <YAxis hide />
-  );
-
-  const tooltipEl = (
-    <Tooltip
-      content={(props) => (
-        <ChartTooltip
-          active={props.active}
-          payload={props.payload}
-          label={props.label}
-          series={series}
-          resolvedColors={resolvedColors}
-        />
-      )}
-      isAnimationActive={!reducedMotion}
-    />
-  );
-
-  const referenceLineEls = effectiveConfig.referenceLines.map((rl, i) => (
-    <ReferenceLine
-      key={i}
-      y={rl.y}
-      x={rl.x}
-      stroke={rl.color ?? 'var(--v-border-default)'}
-      strokeDasharray="4 4"
-      label={rl.label}
-    />
-  ));
-
-  // ─── Chart renderers ──────────────────────────────────────────────────────
-
-  function renderLineChart() {
-    return (
-      <LineChart data={data as ChartDataPoint[]}>
-        {cartesianGridEl}
-        {xAxisEl}
-        {yAxisEl}
-        {tooltipEl}
-        {referenceLineEls}
-        {visibleSeries.map((s) => (
-          <Line
-            key={s.dataKey}
-            type="monotone"
-            dataKey={s.dataKey}
-            name={s.name}
-            stroke={resolvedColors[s.dataKey]}
-            strokeWidth={strokeWidth}
-            dot={{ r: 0 }}
-            activeDot={{ r: ACTIVE_DOT_RADIUS, strokeWidth: 0 }}
-            isAnimationActive={!reducedMotion}
-          />
-        ))}
-      </LineChart>
-    );
-  }
-
-  function renderBarChart() {
-    return (
-      <BarChart data={data as ChartDataPoint[]}>
-        {cartesianGridEl}
-        {xAxisEl}
-        {yAxisEl}
-        {tooltipEl}
-        {referenceLineEls}
-        {visibleSeries.map((s) => (
-          <Bar
-            key={s.dataKey}
-            dataKey={s.dataKey}
-            name={s.name}
-            fill={resolvedColors[s.dataKey]}
-            radius={[BAR_CORNER_RADIUS, BAR_CORNER_RADIUS, 0, 0]}
-            stackId={effectiveConfig.stacked ? 'stack' : undefined}
-            isAnimationActive={!reducedMotion}
-          />
-        ))}
-      </BarChart>
-    );
-  }
-
-  function renderAreaChart() {
-    return (
-      <AreaChart data={data as ChartDataPoint[]}>
-        {cartesianGridEl}
-        {xAxisEl}
-        {yAxisEl}
-        {tooltipEl}
-        {referenceLineEls}
-        {visibleSeries.map((s) => (
-          <Area
-            key={s.dataKey}
-            type="monotone"
-            dataKey={s.dataKey}
-            name={s.name}
-            stroke={resolvedColors[s.dataKey]}
-            fill={resolvedColors[s.dataKey]}
-            fillOpacity={AREA_FILL_OPACITY}
-            strokeWidth={strokeWidth}
-            dot={false}
-            activeDot={{ r: ACTIVE_DOT_RADIUS, strokeWidth: 0 }}
-            stackId={effectiveConfig.stacked ? 'stack' : undefined}
-            isAnimationActive={!reducedMotion}
-          />
-        ))}
-      </AreaChart>
-    );
-  }
-
-  function renderDonutChart() {
-    // Donut uses a single Pie with Cell per series
-    // Data format for donut: each series maps to a pie slice
-    // We aggregate the first numeric value from each series across all data points
-    const donutData = series.map((s) => {
-      const total = data.reduce((sum, d) => {
-        const val = d[s.dataKey];
-        return sum + (typeof val === 'number' ? val : 0);
-      }, 0);
-      return { name: s.name, value: total, dataKey: s.dataKey };
-    });
-
-    const visibleDonutData = donutData.filter((d) => !hiddenSeries.has(d.dataKey));
-    const outerRadius = Math.min(effectiveConfig.height / 2 - 16, 120);
-    const innerRadius = outerRadius * DONUT_INNER_RADIUS_RATIO;
-
-    return (
-      <PieChart>
-        <Pie
-          data={visibleDonutData}
-          cx="50%"
-          cy="50%"
-          innerRadius={innerRadius}
-          outerRadius={outerRadius}
-          dataKey="value"
-          nameKey="name"
-          isAnimationActive={!reducedMotion}
-        >
-          {visibleDonutData.map((entry) => (
-            <Cell
-              key={entry.dataKey}
-              fill={resolvedColors[entry.dataKey]}
-            />
-          ))}
-        </Pie>
-        <Tooltip
-          content={(props) => (
-            <ChartTooltip
-              active={props.active}
-              payload={props.payload}
-              label={props.label}
-              series={series}
-              resolvedColors={resolvedColors}
-            />
-          )}
-          isAnimationActive={!reducedMotion}
-        />
-      </PieChart>
-    );
-  }
-
-  function renderSparklineChart() {
-    // Sparkline: minimal line chart — no axes, legend, or grid
-    const firstSeries = visibleSeries[0] ?? series[0];
-    if (!firstSeries) return null;
-
-    return (
-      <LineChart data={data as ChartDataPoint[]}>
-        <Line
-          type="monotone"
-          dataKey={firstSeries.dataKey}
-          stroke={resolvedColors[firstSeries.dataKey]}
-          strokeWidth={strokeWidth}
-          dot={false}
-          isAnimationActive={!reducedMotion}
-        />
-      </LineChart>
-    );
-  }
-
-  function renderScatterChart() {
-    return (
-      <ScatterChart>
-        <CartesianGrid stroke="var(--v-border-subtle)" strokeDasharray="3 3" />
-        {xAxisEl}
-        {yAxisEl}
-        {tooltipEl}
-        {visibleSeries.map((s) => {
-          const xKey = s.xDataKey ?? effectiveConfig.xAxisKey;
-          // Transform data for scatter: each point needs x and y keys
-          const scatterData = (data as ChartDataPoint[]).map((d) => ({
-            x: d[xKey],
-            y: d[s.dataKey],
-          }));
-
-          return (
-            <Scatter
-              key={s.dataKey}
-              name={s.name}
-              data={scatterData}
-              fill={resolvedColors[s.dataKey]}
-              isAnimationActive={!reducedMotion}
-            />
-          );
-        })}
-      </ScatterChart>
-    );
-  }
-
-  // ─── Chart selector ───────────────────────────────────────────────────────
-
-  function renderChart() {
-    switch (type) {
-      case 'line':
-        return renderLineChart();
-      case 'bar':
-        return renderBarChart();
-      case 'area':
-        return renderAreaChart();
-      case 'donut':
-        return renderDonutChart();
-      case 'sparkline':
-        return renderSparklineChart();
-      case 'scatter':
-        return renderScatterChart();
-      default: {
-        const _exhaustive: never = type;
-        return _exhaustive;
-      }
-    }
-  }
-
   // ─── Layout ───────────────────────────────────────────────────────────────
 
   const isSparkline = type === 'sparkline';
@@ -578,26 +289,26 @@ export function VastuChart({
       aria-label={ariaLabel ?? t('chart.ariaLabel')}
       role="img"
     >
-      {/* ─── Legend top / left ────────────────────────────────────────────── */}
-      {showLegend && (legendPosition === 'top' || legendPosition === 'left') && (
+      {/* ─── Legend top ─────────────────────────────────────────────────────── */}
+      {showLegend && legendPosition === 'top' && (
         <ChartLegend
           series={series}
           resolvedColors={resolvedColors}
           hiddenSeries={hiddenSeries}
-          position={legendPosition}
+          position="top"
           onToggle={handleLegendToggle}
           onSolo={handleLegendSolo}
         />
       )}
 
-      {/* ─── Chart wrapper ────────────────────────────────────────────────── */}
+      {/* ─── Chart wrapper (flex for left/right legends) ──────────────────── */}
       <div
         style={{
           display: isLegendVertical ? 'flex' : 'block',
           flex: 1,
         }}
       >
-        {/* Legend left */}
+        {/* Legend left — inside flex container so it participates in horizontal layout */}
         {showLegend && legendPosition === 'left' && (
           <ChartLegend
             series={series}
@@ -611,7 +322,18 @@ export function VastuChart({
 
         <div className={classes.chartArea} style={{ height: chartHeight }}>
           <ResponsiveContainer width="100%" height="100%">
-            {renderChart() ?? <div />}
+            <ChartRenderer
+              type={type}
+              data={data}
+              visibleSeries={visibleSeries}
+              allSeries={series}
+              hiddenSeries={hiddenSeries}
+              resolvedColors={resolvedColors}
+              effectiveConfig={effectiveConfig}
+              strokeWidth={strokeWidth}
+              reducedMotion={reducedMotion}
+              chartElements={chartElements}
+            />
           </ResponsiveContainer>
 
           {/* Config gear button — shown unless sparkline */}
@@ -652,7 +374,7 @@ export function VastuChart({
         )}
       </div>
 
-      {/* ─── Legend bottom ────────────────────────────────────────────────── */}
+      {/* ─── Legend bottom ──────────────────────────────────────────────────── */}
       {showLegend && legendPosition === 'bottom' && (
         <ChartLegend
           series={series}
